+++ src/StudioCore/MapStudioNew.cs
@@ -25,7 +25,7 @@ using System.Numerics;
 using System.Runtime.InteropServices;
 using Veldrid;
 using Veldrid.Sdl2;
-using StudioCore.Interface;
+using StudioCore.Editor;
 using StudioCore.Utilities;
 using Renderer = StudioCore.Scene.Renderer;
 using Thread = System.Threading.Thread;
